moved Cart to function
