+++ index.js
@@ -6,12 +6,10 @@ const morgan = require('koa-morgan');
 const render = require('koa-ejs');
 const KoaRouter = require('koa-router');
 const koaBody = require('koa-bodyparser');
-const chalk = require('chalk');
-const debug = require('debug')('index');
 const path = require('path');
 const rfs = require('rotating-file-stream');
 const { articlesFromPage, getArticle } = require('./articlesFromPage');
-const getPrice = require('./getPrice');
+const getCart = require('./getPrice');
 
 const app = new Koa();
 const port = process.env.PORT || 3000;
@@ -34,95 +32,19 @@ router.get('/', ctx => {
 });
 
 router.post('/recipes/search/:page', async ctx => {
-  debug(`the request body is ${chalk.green(ctx.request.body)}`);
   const whatToSearch = await articlesFromPage(ctx.request.body.ings, ctx.params.page);
-  debug(ctx.params.page);
   return ctx.render('searchResults', { recipesArray: whatToSearch });
 });
 
 router.get('/recipe/:id', async ctx => {
-  debug(ctx.params.id);
   const article = await getArticle(`https://www.povarenok.ru/recipes/show/${ctx.params.id}`);
-  debug(article);
   // TO DO
   return ctx.render('reciepPage', article);
 });
 router.post('/Cart', async ctx => {
-  let listOfProducts;
-  // eslint-disable-next-line no-restricted-syntax
-  for (const product in ctx.request.body) {
-    if (Object.prototype.hasOwnProperty.call(ctx.request.body, product)) {
-      listOfProducts = ctx.request.body[product];
-    }
-  }
-
-  let arrayPrice;
-  const arrayOfStore = [];
-  arrayOfStore[0] = {
-    name: 'Novus',
-    price: 0,
-    count: 0,
-    products: [...listOfProducts],
-    MinimalCost: false
-  };
-  arrayOfStore[1] = {
-    name: 'АТБ',
-    price: 0,
-    count: 0,
-    products: [...listOfProducts],
-    MinimalCost: false
-  };
-  arrayOfStore[2] = {
-    name: 'Велика Кишеня',
-    price: 0,
-    count: 0,
-    products: [...listOfProducts],
-    MinimalCost: false
-  };
-  arrayOfStore[3] = {
-    name: 'Сільпо',
-    price: 0,
-    count: 0,
-    products: [...listOfProducts],
-    MinimalCost: false
-  };
-  for (let i = 0; i < listOfProducts.length; i++) {
-    // eslint-disable-next-line no-await-in-loop
-    arrayPrice = await getPrice([listOfProducts[i]]);
-    const object = arrayPrice[0].pricesAndStores;
-
-    //  eslint-disable-next-line guard-for-in
-    for (const key in object) {
-      for (let store = 0; store < arrayOfStore.length; store++) {
-        if (arrayOfStore[store].name === object[key].store) {
-          arrayOfStore[store].price += object[key].price;
-
-          arrayOfStore[store].count += 1;
-
-          for (let product = 0; product < arrayOfStore[store].products.length; product++) {
-            if (listOfProducts[i] === arrayOfStore[store].products[product]) {
-              arrayOfStore[store].products.splice(product, 1);
-            }
-          }
-        }
-      }
-    }
-  }
-  let minStore = arrayOfStore[0];
-  const countFirst = arrayOfStore[0].count;
-  for (let i = 1; i < arrayOfStore.length; i++) {
-    if (countFirst === arrayOfStore[i].count) {
-      if (minStore.price > arrayOfStore[i].price) {
-        minStore = arrayOfStore[i];
-      }
-    } else if (countFirst < arrayOfStore[i].count) {
-      minStore = arrayOfStore[i];
-    }
-  }
-  for (let i = 0; i < arrayOfStore.length; i++)
-    if (arrayOfStore[i].name === minStore.name) arrayOfStore[i].MinimalCost = true;
-    console.log(arrayOfStore);
-  return ctx.body = arrayOfStore;
+  const result = await getCart(ctx.request.body);
+  // eslint-disable-next-line no-return-assign
+  return (ctx.body = result);
 });
 app
   .use(koaBody())
@@ -134,6 +56,4 @@ app
     ctx.body = ctx.request.body;
   });
 
-app.listen(port, () => {
-  debug(`listening on port ${chalk.green('3000')}`);
-});
+app.listen(port);
